feat: skeleton preview for SimpleItinerary; modify: the duration for the skeleton preview
